makefile

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: sinseungheon <sinseungheon@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/26 23:21:25 by sinseungheo       #+#    #+#             */
-/*   Updated: 2025/07/03 23:46:18 by sinseungheo      ###   ########.fr       */
+/*   Updated: 2025/07/09 00:08:53 by sinseungheo      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *restrict dst, const void *restrict src, size_t n)
 {
-	int i;
+	size_t i;
 	unsigned char *d;
 	const unsigned char *s;
