+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: sinseungheon <sinseungheon@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/03 22:55:37 by sinseungheo       #+#    #+#             */
-/*   Updated: 2025/07/03 23:46:37 by sinseungheo      ###   ########.fr       */
+/*   Updated: 2025/07/09 00:10:08 by sinseungheo      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t ft_strlcat(char *restrict dst, const char *restrict src, size_t dstsize)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (dst[i] != 0)
