+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: sinseungheon <sinseungheon@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/05 15:50:35 by sinseungheo       #+#    #+#             */
-/*   Updated: 2025/07/05 17:07:21 by sinseungheo      ###   ########.fr       */
+/*   Updated: 2025/07/09 00:15:57 by sinseungheo      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int		i;
+	size_t	i;
 	int		k;
 	char	*strh;
 	char	*strn;
